Switched to the legacy version of libarchive.
Since we're not really using anything special from the new version, it makes a bit of sense to use the
legacy version. It compiles to be a lot smaller than the stable version. This library is 475K with
compiler optimizations enabled for size. The stable version was 1.7M without optimizations (and 1M with).
If there is a compelling reason to use the stable version in the future, we will.